+++ valpas-web/src/state/koodistot.ts
@@ -1,6 +1,9 @@
 import { LocalizedString } from "./types"
 
-export type KoodistoKoodiviite<T extends string, S extends string> = {
+export type KoodistoKoodiviite<
+  T extends string = string,
+  S extends string = string
+> = {
   koodistoUri: T
   koodiarvo: S
   koodistoVersio?: number
@@ -8,56 +11,36 @@ export type KoodistoKoodiviite<T extends string, S extends string> = {
   lyhytNimi?: LocalizedString
 }
 
-export type Opiskeluoikeudentyyppi = KoodistoKoodiviite<
-  "opiskeluoikeudentyyppi",
-  string
->
+const koodiarvoCondition = <K extends KoodistoKoodiviite, S extends string>(
+  uri: string,
+  truthyValues: S[]
+) => (koodiviite: K) =>
+  uri === koodiviite.koodistoUri &&
+  truthyValues.includes(koodiviite.koodiarvo as S)
 
-export type OppilaitosnumeroKoodistoviite = KoodistoKoodiviite<
-  "oppilaitosnumero",
-  string
->
+export type Opiskeluoikeudentyyppi = KoodistoKoodiviite<"opiskeluoikeudentyyppi">
 
-export type PaikkakuntaKoodistoviite = KoodistoKoodiviite<"paikkakunta", string>
-
+// TODO: Omakeksittyjä koodistoja. Pitää tsekata mitä löytyy valmiina.
+export type OppilaitosnumeroKoodistoviite = KoodistoKoodiviite<"oppilaitosnumero">
+export type PaikkakuntaKoodistoviite = KoodistoKoodiviite<"paikkakunta">
 export type HakemuksentilaKoodistoviite = KoodistoKoodiviite<
   "hakemuksentila",
   "aktiivinen" | "passiivinen" | "puutteellinen" | "luonnos"
 >
-
 export type ValintatietotilaKoodistoviite = KoodistoKoodiviite<
   "valintatietotila",
   "hyväksytty" | "varasija" | "hylätty" | "vastaanotettu" | "läsnä"
 >
 
-export const isHyväksytty = (
-  valintatietotila: ValintatietotilaKoodistoviite
-): boolean => {
-  const arvo = valintatietotila.koodiarvo
-  return arvo === "hyväksytty" || arvo === "vastaanotettu" || arvo === "läsnä"
+export const ValintatietotilaKoodistoviite = {
+  isHyväksytty: koodiarvoCondition("valintatietotila", [
+    "hyväksytty",
+    "vastaanotettu",
+    "läsnä",
+  ]),
+  isVastaanotettu: koodiarvoCondition("valintatietotila", [
+    "vastaanotettu",
+    "läsnä",
+  ]),
+  isLäsnä: koodiarvoCondition("valintatietotila", ["läsnä"]),
 }
-
-export const isVastaanotettu = (
-  valintatietotila: ValintatietotilaKoodistoviite
-): boolean => {
-  const arvo = valintatietotila.koodiarvo
-  return arvo === "vastaanotettu" || arvo === "läsnä"
-}
-
-export const isLäsnä = (
-  valintatietotila: ValintatietotilaKoodistoviite
-): boolean => {
-  const arvo = valintatietotila.koodiarvo
-  return arvo === "läsnä"
-}
-
-export const mockKoodistoviite = <T extends string>(koodistoUri: T) => <
-  S extends string
->(
-  koodiarvo: S,
-  nimi?: string
-): KoodistoKoodiviite<T, S> => ({
-  koodistoUri,
-  koodiarvo,
-  nimi: { fi: `${nimi || koodiarvo}` },
-})
