+++ valpas-web/src/state/types.ts
@@ -3,8 +3,12 @@ import {
   PaikkakuntaKoodistoviite,
 } from "./koodistot"
 
-export type Oid = `1.${number}.${number}.${number}.${number}.${number}.${number}`
-export type ISODate = `${number}-${number}-${number}`
+// Tempate literal -tyypitykset aiheuttavat stack overflow'n linttausvaiheessa.
+// Korjattaneen Typescriptin versiossa 4.2, väliaikaisesti mennään yksinkertaisemmalla tyypityksellä.
+// export type Oid = `1.${number}.${number}.${number}.${number}.${number}.${number}`
+// export type ISODate = `${number}-${number}-${number}`
+export type Oid = string
+export type ISODate = string
 
 export type Language = "fi" | "sv" | "en"
 export type LocalizedString = Partial<Record<Language, string>>
